airflow_code/requirements.txt
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.2
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
appdirs==1.4.3
argcomplete==2.0.0
attrs==20.3.0
Automat==0.8.0
Babel==2.10.1
blinker==1.4
botocore==1.29.27
cachelib==0.7.0
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
cloud-init==22.1
colorama==0.4.4
colorlog==4.8.0
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
connexion==2.13.1
constantly==15.1.0
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
dbus-python==1.2.16
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.14.0
httpx==0.23.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.10.1
incremental==16.10.1
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.5.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==21.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.6.0
pluggy==1.0.0
prison==0.2.1
psutil==5.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
Pygments==2.12.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-daemon==2.3.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==12.4.4
s3transfer==0.6.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.2.3
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sos==4.3
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
ssh-import-id==5.10
swagger-ui-bundle==0.0.9
systemd-python==234
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
Twisted==18.9.0
typing-extensions==4.2.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.26.9
virtualenv==20.0.17
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
zope.interface==4.7.1
